chore: Refactor nav.html to improve loading performance and enable financial app
Refactor the nav.html file to improve the loading performance by lazy loading the images. Additionally, enable the financial app by loading the financial data and setting the `enabledApps.financial` flag to true. This change also fixes a bug where the financial app was not being loaded if the user profile was not defined.

Closes #123